Refactor FetchPlanState to org.datanucleus. Fixes #149
