+++ src/main/java/org/datanucleus/ExecutionContextImpl.java
@@ -75,7 +75,6 @@ import org.datanucleus.metadata.TransactionType;
 import org.datanucleus.properties.BasePropertyStore;
 import org.datanucleus.state.CallbackHandler;
 import org.datanucleus.state.DetachState;
-import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.LockManager;
 import org.datanucleus.state.LockManagerImpl;
 import org.datanucleus.state.NullCallbackHandler;
