+++ src/main/java/org/datanucleus/ExecutionContextThreadedImpl.java
@@ -22,7 +22,6 @@ import java.util.Map;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
-import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
 import org.datanucleus.store.Extent;
