+++ src/main/java/org/datanucleus/state/AbstractStateManager.java
@@ -24,6 +24,7 @@ import java.util.concurrent.locks.ReentrantLock;
 
 import org.datanucleus.ExecutionContext;
 import org.datanucleus.FetchPlanForClass;
+import org.datanucleus.FetchPlanState;
 import org.datanucleus.PropertyNames;
 import org.datanucleus.cache.CachedPC;
 import org.datanucleus.cache.L2CachePopulateFieldManager;
