+++ src/main/java/org/datanucleus/store/fieldmanager/DetachFieldManager.java
@@ -24,12 +24,12 @@ import java.util.Map.Entry;
 
 import org.datanucleus.ExecutionContext;
 import org.datanucleus.FetchPlanForClass;
+import org.datanucleus.FetchPlanState;
 import org.datanucleus.api.ApiAdapter;
 import org.datanucleus.metadata.AbstractMemberMetaData;
 import org.datanucleus.metadata.MapMetaData;
 import org.datanucleus.metadata.RelationType;
 import org.datanucleus.state.DetachState;
-import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
 import org.datanucleus.store.types.ContainerHandler;
 import org.datanucleus.store.types.ElementContainerAdapter;
