+++ src/main/java/org/datanucleus/store/fieldmanager/LoadFieldManager.java
@@ -20,10 +20,10 @@ package org.datanucleus.store.fieldmanager;
 
 import org.datanucleus.ExecutionContext;
 import org.datanucleus.FetchPlanForClass;
+import org.datanucleus.FetchPlanState;
 import org.datanucleus.api.ApiAdapter;
 import org.datanucleus.metadata.AbstractMemberMetaData;
 import org.datanucleus.metadata.RelationType;
-import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
 import org.datanucleus.store.types.ContainerHandler;
 import org.datanucleus.store.types.SCO;
