+++ src/main/java/org/datanucleus/store/types/wrappers/GregorianCalendar.java
@@ -22,8 +22,8 @@ import java.io.ObjectStreamException;
 import java.util.Date;
 import java.util.TimeZone;
 
+import org.datanucleus.FetchPlanState;
 import org.datanucleus.metadata.AbstractMemberMetaData;
-import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
 import org.datanucleus.store.types.SCO;
