+++ src/main/java/org/datanucleus/store/types/wrappers/TreeMap.java
@@ -23,10 +23,10 @@ import java.util.Comparator;
 import java.util.Iterator;
 import java.util.SortedMap;
 
+import org.datanucleus.FetchPlanState;
 import org.datanucleus.flush.MapPutOperation;
 import org.datanucleus.flush.MapRemoveOperation;
 import org.datanucleus.metadata.AbstractMemberMetaData;
-import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
 import org.datanucleus.store.types.SCOMap;
 import org.datanucleus.store.types.SCOUtils;
